Drop support for Cura 4.5 and earlier.
Since there is more work to do in order to support correctly
the print profiles, and they add value to the integration
implemented in order to better work with the Finder, I'm
dropping support now until I can better test and detect
older versions of Cura.

Later versions can upgrade the profiles based on the settings
versions and as such should not cause any issues.