bugfix: fix temperature with a floating point.
The start gcode macros in recent versions of Cura
format the print temperature as a floating point.
Other temperature commands in the gcode don't have
such behavior.

On Finder (version 1?) with the more recent firmware, this
causes the temperature to be the fixed to start at 185C,
which appears to be the "default start temperature" or some
rounding error when reading the floating point temperature.

This commit fixes that by parsing the 'M104 SXXX' commands,
and changing the temperature by converting it to integer.
This also allows the temperature to be embedded to the .gx
file header.

I took the opportunity to move the parsing commands into the
gx.py script, as a way to potentially easy the proccess to
make this a standalone script that can potentially be used
by other slicers/software.

Fixes #21
Related to #2 and #16